% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{plot.kmeans}
\alias{plot.kmeans}
\title{Plot Method for K-Means Fits}
\usage{
\method{plot}{kmeans}(fit, boxplot = F)
}
\arguments{
\item{fit}{an object of class "kmeans", usually, a result of a call to kmeans}

\item{boxplot}{whether to plot box-and-whisker around data points}
}
\description{
summary method for class "kmeans".
}
