% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{search.k}
\alias{search.k}
\title{Find best k for K-Means Clustering}
\usage{
search.k(x, Ks, nstart = 100)
}
\arguments{
\item{x}{numeric matrix of data, or an object that can be coerced to such a matrix (such as a numeric vector or a data frame with all numeric columns).}

\item{Ks}{a vector of candidate K values to perform K-Means Clustering}

\item{nstart}{number of times to repeat with different random seeds for each k}
}
\description{
Search for the best k to perform k-means clustering on a data matrix.
}
