% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{summary.kmeans}
\alias{summary.kmeans}
\title{Summarizing K-Means Fits}
\usage{
\method{summary}{kmeans}(fit)
}
\arguments{
\item{fit}{an object of class "kmeans", usually, a result of a call to kmeans}
}
\description{
summary method for class "kmeans".
}
